4e00c477 fix redeclared without dllimport comp warnings
1f50ef3f address size_t to int type conversion compilation warnings
c0172288 fix set but not used warning

+++ Filters/Core/vtkDataObjectGenerator.cxx
@@ -703,7 +703,7 @@ vtkDataObject * vtkDataObjectGenerator::FillOutputDataObjects(
       }
 
     double origin[3] = {0,0,0};
-    hbo->Initialize(blocksPerLevel.size(), &blocksPerLevel[0], origin, VTK_XYZ_GRID);
+    hbo->Initialize(static_cast<int>(blocksPerLevel.size()), &blocksPerLevel[0], origin, VTK_XYZ_GRID);
     vtkIdType gcnt = 0;
     for (git = structure->children.begin();
          git != structure->children.end();
